+++ core/geometry.py
