+++ label_tools/kitti_object/kitti_object_data_loader.py
@@ -9,7 +9,7 @@ import numpy as np
 from pathlib import Path
 
 sys.path.append(Path(__file__).parent.parent.as_posix())
-from utils.transform import *
+from core.transform import *
 
 
 def load_lidar_data(path: str):
