+++ label_tools/kitti_object/kitti_object_helper.py
@@ -9,9 +9,9 @@ from pathlib import Path
 import transforms3d.euler
 
 sys.path.append(Path(__file__).parent.parent.as_posix())
-from utils.geometry_types import Transform, Location
-from utils.label_types import ObjectLabel
-from utils.transform import bbox_to_o3d_bbox
+from core.geometry import Transform, Location
+from core.types import ObjectLabel
+from core.transform import bbox_to_o3d_bbox
 
 
 class Param:
