+++ recorder/actor.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python3
 import carla
-from utils.transform import *
+from core.transform import *
 
 
 class PseudoActor(object):
