+++ recorder/actor_factory.py
@@ -10,8 +10,8 @@ import yaml
 import carla
 
 from param import RAW_DATA_PATH, ROOT_PATH
-from utils.geometry_types import *
-from utils.transform import transform_to_carla_transform
+from core.geometry import *
+from core.transform import transform_to_carla_transform
 
 from recorder.actor import Actor, PseudoActor
 from recorder.camera import RgbCamera, DepthCamera, SemanticSegmentationCamera
