+++ recorder/camera.py
@@ -7,8 +7,8 @@ import transforms3d
 import math
 
 from recorder.sensor import Sensor
-from utils.geometry_types import Transform, Rotation
-from utils.transform import carla_transform_to_transform
+from core.geometry import Transform, Rotation
+from core.transform import carla_transform_to_transform
 
 
 class CameraBase(Sensor):
