+++ recorder/world.py
@@ -6,8 +6,8 @@ import carla
 from dataclasses import dataclass
 
 from recorder.actor import PseudoActor
-from utils.label_types import *
-from utils.transform import carla_bbox_to_bbox, carla_transform_to_transform
+from core.types import *
+from core.transform import carla_bbox_to_bbox, carla_transform_to_transform
 
 # Get logger instance
 logger = logging.getLogger(__name__)
