+++ tests/test_logger.py
@@ -20,7 +20,7 @@ from io import StringIO
 # Add parent directory to path for imports
 sys.path.insert(0, str(Path(__file__).parent.parent))
 
-from utils.logger import (
+from core.logger import (
     configure_global_logging,
     get_logger,
     setup_logger,
