+++ tools/__init__.py
@@ -1,0 +1,10 @@
+"""
+CARLA Dataset Tools - CLI Tools
+
+This directory contains standalone CLI tools organized by function prefix:
+- viz_*: Visualization tools (lidar, map)
+- editor_*: Interactive editors (route)
+- data_*: Data processing tools (imageset generation)
+- config_*: Configuration management (convert, list, validate)
+- debug_*: Debugging utilities
+"""
