+++ tools/debug_info.py
@@ -4,7 +4,7 @@ import carla
 import sys
 from pathlib import Path
 sys.path.append(Path(__file__).parent.parent.as_posix())
-from utils.transform import carla_transform_to_transform
+from core.transform import carla_transform_to_transform
 
 
 def show_spawn_points(world: carla.World, debug_helper: carla.DebugHelper):
